v5.4.0 fixed test cases
rolled back ffguess control in C13_ftbl;
fixed pooling metabs for _i case;
fixed scale factors when many types of measurements are available;
updated test/ok files
added cases in test/
added test/README.txt

+++ clean_ftbl.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # clean files after runing influx_s[i] in the current (default) or a directory given in first and next parameters
 
 if [ $# != 0 ]; then
@@ -8,9 +8,11 @@ else
 fi
 
 for d in $dirs; do
+   [ -e "$d" ] || continue
    ( cd "$d" &&
    for f in *.ftbl; do
       #echo fb="${f/.ftbl/}"
+      #echo "f="$f
       fb="${f/.ftbl/}"
       rm -f "$fb"{.R,.err,.log,.pres.csv,.RData,.Rprof,_res.kvh,_ires.kvh,.netan,.sif,.xgmml,_netan.kvh,_rev.txt,_net.txt,_fwd.txt,.xml,.pdf} *.$fb.attrs {edge,node}.*.$fb*\(.V+\([0-9]\)\) "$fb".mc*.err parallel.R
    done )
